# docs/swagger.yaml
basePath: /api
definitions:
...
  controllers.BulkDeleteDBPolicyRequest:
    properties:
      policy_ids:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - policy_ids
    type: object
  controllers.BulkDeleteDBPolicyResponse:
    properties:
      deleted_count:
        type: integer
      errors:
        items:
          type: string
        type: array
      failed_ids:
        items:
          type: integer
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  controllers.BulkPolicyUpdateJobResponse:
    properties:
      message:
        example: 'Bulk policy update background job started: F.C4LG5UMHSAPEI0J. Adding
          6 policies, removing 3 policies.'
        type: string
      status:
        example: job_started
        type: string
    type: object
  controllers.BulkPolicyUpdateRequest:
    properties:
      cntmgt_id:
        example: 1
        type: integer
      dbactormgt_id:
        example: 3
        type: integer
      dbmgt_id:
        example: 5
        type: integer
      new_object_mgts:
        example:
        - 100
        - 101
        - 102
        items:
          type: integer
        type: array
      new_policy_defaults:
        example:
        - 10
        - 11
        - 12
        items:
          type: integer
        type: array
    required:
    - cntmgt_id
    - dbactormgt_id
    - dbmgt_id
    - new_object_mgts
    - new_policy_defaults
    type: object
...
  controllers.DBPolicyCreateRequest:
    properties:
      cntmgt:
        example: 1
        type: integer
      dbactormgt:
        example: 1
        type: integer
      dbmgt:
        example: 1
        type: integer
      dbobjectmgt:
        example: 1
        type: integer
      dbpolicydefault:
        example: 1
        type: integer
      description:
        example: Policy description
        type: string
      status:
        example: active
        type: string
    type: object
  controllers.DBPolicyCreateResponse:
    properties:
      id:
        example: 123
        type: integer
      message:
        example: Policy created
        type: string
    type: object
  controllers.DBPolicyDeleteResponse:
    properties:
      message:
        example: Policy deleted
        type: string
    type: object
  controllers.DBPolicyUpdateResponse:
    properties:
      message:
        example: Policy updated
        type: string
    type: object
...
  controllers.DownloadErrorResponse:
    properties:
      error:
        example: internal server error
        type: string
      message:
        example: failed to execute filedownload
        type: string
    type: object
  controllers.DownloadJobStartResponse:
    properties:
      job_id:
        example: filedownload_config_1234567890
        type: string
      message:
        example: 'Download job started successfully for file: config.json'
        type: string
      status:
        example: success
        type: string
    type: object
  controllers.DownloadValidationErrorResponse:
    properties:
      error:
        example: validation error
        type: string
      message:
        example: source_path is required
        type: string
    type: object
...
  controllers.InternalServerErrorResponse:
    properties:
      error:
        example: database connection failed
        type: string
    type: object
...
  controllers.JobProcessingErrorResponse:
    properties:
      error:
        example: failed to start background job
        type: string
    type: object
  controllers.JobStartResponse:
    properties:
      message:
        example: 'Background job started: job_12345'
        type: string
    type: object
...
  controllers.NotFoundResponse:
    properties:
      error:
        example: record not found
        type: string
    type: object
...
  controllers.PolicyCreationErrorResponse:
    properties:
      error:
        example: failed to create policy
        type: string
    type: object
...
  controllers.StandardErrorResponse:
    properties:
      error:
        example: Invalid request body
        type: string
    type: object
...
  models.DownloadRequest:
    properties:
      cnt_id:
        example: 456
        type: integer
      save_path:
        example: /root/
        type: string
      source_path:
        example: /var/data/config.json
        type: string
    required:
    - cnt_id
    - save_path
    - source_path
    type: object
...
info:
  contact: {}
  description: DBF Artifact API
  title: dbfartifactapi
  version: "1.0"
paths:
...
  /api/queries/dbpolicy:
    post:
      consumes:
      - application/json
      description: Creates a new database policy with specified parameters
      parameters:
      - description: Database Policy object
        in: body
        name: policy
        required: true
        schema:
          $ref: '#/definitions/controllers.DBPolicyCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Policy created successfully with ID
          schema:
            $ref: '#/definitions/controllers.DBPolicyCreateResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/controllers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.PolicyCreationErrorResponse'
      summary: Create database policy
      tags:
      - DB Policy
  /api/queries/dbpolicy/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an existing database policy by ID. Revokes permissions
        via SqlUpdateDeny before removing policy record.
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Policy deleted successfully
          schema:
            $ref: '#/definitions/controllers.DBPolicyDeleteResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/controllers.StandardErrorResponse'
        "404":
          description: Policy not found
          schema:
            $ref: '#/definitions/controllers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.InternalServerErrorResponse'
      summary: Delete database policy
      tags:
      - DB Policy
    put:
      consumes:
      - application/json
      description: Updates an existing database policy by ID
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Database Policy object
        in: body
        name: policy
        required: true
        schema:
          $ref: '#/definitions/controllers.DBPolicyCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Policy updated successfully
          schema:
            $ref: '#/definitions/controllers.DBPolicyUpdateResponse'
        "400":
          description: Invalid ID or request body
          schema:
            $ref: '#/definitions/controllers.StandardErrorResponse'
        "404":
          description: Policy not found
          schema:
            $ref: '#/definitions/controllers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.InternalServerErrorResponse'
      summary: Update database policy
      tags:
      - DB Policy
  /api/queries/dbpolicy/bulkdelete:
    post:
      consumes:
      - application/json
      description: Deletes multiple database policies by IDs. Revokes permissions
        via SqlUpdateDeny for enabled policies before removing records. Returns count
        of successfully deleted policies and list of failed IDs if any.
      parameters:
      - description: Policy IDs to delete
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.BulkDeleteDBPolicyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Bulk delete completed with success/failure details
          schema:
            $ref: '#/definitions/controllers.BulkDeleteDBPolicyResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/controllers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.InternalServerErrorResponse'
      summary: Bulk delete database policies
      tags:
      - DB Policy
  /api/queries/dbpolicy/bulkupdate:
    post:
      consumes:
      - application/json
      description: Compares existing policies with desired state (Cartesian product
        of policy_defaults × objects) and executes changes via VeloArtifact background
        job. Only updates database after successful remote execution to ensure atomic
        consistency.
      parameters:
      - description: Bulk policy update request with actor, policy defaults, and objects
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.BulkPolicyUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Background job started with job ID and change summary
          schema:
            $ref: '#/definitions/controllers.BulkPolicyUpdateJobResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/controllers.StandardErrorResponse'
        "404":
          description: Referenced entity (cntmgt, dbmgt, actor, policy, object) not
            found
          schema:
            $ref: '#/definitions/controllers.NotFoundResponse'
        "500":
          description: Internal server error during job creation or VeloArtifact execution
          schema:
            $ref: '#/definitions/controllers.InternalServerErrorResponse'
      summary: Bulk update database policies
      tags:
      - DB Policy
  /api/queries/dbpolicy/cntmgt/{cntmgt}:
    get:
      consumes:
      - application/json
      description: Starts background job to generate database policies for all databases
        under specified connection management ID
      parameters:
      - description: Connection Management ID
        in: path
        name: cntmgt
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Background job started message with job ID
          schema:
            $ref: '#/definitions/controllers.JobStartResponse'
        "400":
          description: Invalid connection management ID
          schema:
            $ref: '#/definitions/controllers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.JobProcessingErrorResponse'
      summary: Generate policies for connection management
      tags:
      - DB Policy
...
  /api/queries/download:
    post:
      consumes:
      - application/json
      description: Starts background job to download file from server to agent. If
        source_path is a directory, it will be compressed to tar.gz before sending.
        The agent will verify MD5 hash and extract compressed files automatically.
      parameters:
      - description: Download request with connection ID, source path, and save path
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DownloadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Download job started successfully with job ID
          schema:
            $ref: '#/definitions/controllers.DownloadJobStartResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/controllers.DownloadValidationErrorResponse'
        "500":
          description: Internal server error during download execution
          schema:
            $ref: '#/definitions/controllers.DownloadErrorResponse'
      summary: Execute file download operation
      tags:
      - Download
...
swagger: "2.0"